pyproject.toml
[project]
name = "analytics-ai-agent-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-bigquery>=3.32.0",
    "pandas>=2.2.3",
    "plotly[express]>=6.1.0",
    "pydantic-ai>=0.2.4",
]

[dependency-groups]
dev = [
    "ruff>=0.11.10",
]
